Refactor Form component to encapsulate its state
