+++ frontend/src/components/Form.tsx
@@ -1,14 +1,13 @@
-import * as React from "react";
+import React, { useEffect, useState } from "react";
 import Grid from "@mui/material/Grid";
-import {
-  Control,
-  Controller,
-  UseFormHandleSubmit,
-  UseFormSetValue,
-} from "react-hook-form";
+import { Controller, useForm } from "react-hook-form";
+import axios from "axios";
+import { yupResolver } from "@hookform/resolvers/yup";
+import * as yup from "yup";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import Switch from "@mui/material/Switch";
 import Button from "@mui/material/Button";
+import { CalcSpectrumPlotData, CalcSpectrumResponseData } from "../constants";
 import { Database } from "./fields/Database";
 import { Mode } from "./fields/Mode";
 import { TGas } from "./fields/TGas";
@@ -22,35 +21,245 @@ import { WavenumberRangeSlider } from "./fields/WavenumberRangeSlider";
 import { CalcSpectrumButton } from "./fields/CalSpectrumButton";
 import { FormValues } from "./types";
 
+export interface Response<T> {
+  data?: T;
+  error?: string;
+}
+
 interface FormProps {
-  control: Control<FormValues, object>;
-  setValue: UseFormSetValue<FormValues>;
-  handleSubmit: UseFormHandleSubmit<FormValues>;
-  downloadButton: boolean;
-  isNonEquilibrium: boolean;
-  setIsNonEquilibrium: React.Dispatch<React.SetStateAction<boolean>>;
-  useSlit: boolean;
-  setUseSlit: React.Dispatch<React.SetStateAction<boolean>>;
-  useSimulateSlitFunction: boolean;
-  useGesia: boolean;
-  useHitemp: boolean;
-  onSubmit: (data: FormValues, endpoint: string) => Promise<void>;
+  setPlotData: React.Dispatch<
+    React.SetStateAction<CalcSpectrumPlotData | undefined>
+  >;
+  setError: React.Dispatch<React.SetStateAction<string | undefined>>;
+  setLoading: React.Dispatch<React.SetStateAction<boolean>>;
+  setProgress: React.Dispatch<React.SetStateAction<number>>;
+  setCalcSpectrumResponse: React.Dispatch<
+    React.SetStateAction<Response<CalcSpectrumResponseData> | undefined>
+  >;
 }
 
 export const Form: React.FunctionComponent<FormProps> = ({
-  control,
-  setValue,
-  handleSubmit,
-  downloadButton,
-  isNonEquilibrium,
-  setIsNonEquilibrium,
-  useSlit,
-  setUseSlit,
-  useSimulateSlitFunction,
-  useGesia,
-  useHitemp,
-  onSubmit,
+  setPlotData,
+  setError,
+  setLoading,
+  setProgress,
+  setCalcSpectrumResponse,
 }) => {
+  const [isNonEquilibrium, setIsNonEquilibrium] = useState(false);
+  const [useGesia, setUseGesia] = useState(false);
+  const [useHitemp, setUseHitemp] = useState<boolean>(false); //hitemp
+  const [useSlit, setUseSlit] = useState(false); // checking that user wants to apply the slit function or not in available modes
+  const [useSimulateSlitFunction, setUseSimulateSlitFunction] = useState(false); // checking the mode and enable or disable slit feature
+  const [downloadButton, setDownloadButton] = useState(true);
+
+  useEffect(() => {
+    if (useGesia) {
+      // GESIA does not work for non-equilibrium calculations
+      setIsNonEquilibrium(false);
+    }
+  }, [useGesia]);
+
+  const Schema = yup.object().shape({
+    useNonEqi: yup.boolean(),
+    use_simulate_slit: yup.boolean(),
+    path_length: yup
+      .number()
+      .required("Path length must be defined")
+      .typeError("Path length must be defined")
+      .min(1, "Path length cannot be negative"),
+    pressure: yup
+      .number()
+      .required("Pressure must be defined")
+      .typeError("Pressure must be defined")
+      .min(1, "Pressure cannot be negative"),
+    tgas: yup
+      .number()
+      .required("Tgas must be defined")
+      .typeError("Tgas must be defined")
+      .max(9000, "Tgas must be between 1K and 9000K")
+      .min(1, "Tgas must be between 1K and 9000K"),
+    trot: yup
+      .number()
+      .typeError("TRot must be defined")
+      .when("useNonEqi", {
+        is: true,
+        then: yup
+          .number()
+          .required("Trot must be defined")
+          .typeError("TRot must be defined")
+          .min(0, "TRot must be positive"),
+      }),
+    tvib: yup
+      .number()
+      .typeError("TRot must be defined")
+      .when("useNonEqi", {
+        is: true,
+        then: yup
+          .number()
+          .required("TVib must be defined")
+          .typeError("TVib must be defined")
+          .min(0, "TVib must be positive"),
+      }),
+    min_wavenumber_range: yup
+      .number()
+      .required("Min wavenumber range must be defined")
+      .typeError("Min wavenumber range must be defined"),
+    max_wavenumber_range: yup
+      .number()
+      .required("Max wavenumber range must be defined")
+      .typeError("Max wavenumber range must be defined"),
+    species: yup.array().of(
+      yup.object().shape({
+        molecule: yup
+          .string()
+          .required("Molecule must be defined")
+          .typeError("Molecule must be defined"),
+        mole_fraction: yup
+          .number()
+          .required("Mole fraction must be defined")
+          .typeError("Mole fraction must be defined"),
+      })
+    ),
+    simulate_slit: yup
+      .number()
+      .typeError("Simulate slit must be defined")
+      .min(0, "Simulate slit must be positive")
+      .max(30, "Simulate slit must be less than 30")
+      .when("useSlitSwitch", {
+        is: true,
+        then: yup
+          .number()
+          .typeError("Simulate slit must be defined")
+          .min(0, "Simulate slit must be positive")
+          .max(30, "Simulate slit must be less than 30"),
+      }),
+  });
+  const { control, handleSubmit, setValue, watch } = useForm<FormValues>({
+    defaultValues: { species: [{ molecule: "CO", mole_fraction: 0.1 }] },
+    resolver: yupResolver(Schema),
+  });
+
+  const databaseWatch = watch("database");
+  const modeWatch = watch("mode");
+  React.useEffect(() => {
+    if (databaseWatch === "geisa") {
+      setUseGesia(true);
+    } else {
+      setUseGesia(false);
+    }
+    if (databaseWatch === "hitemp") {
+      setUseHitemp(true);
+    } else {
+      setUseHitemp(false);
+    }
+    if (modeWatch === "absorbance") {
+      setUseSimulateSlitFunction(false);
+    } else {
+      setUseSimulateSlitFunction(true);
+    }
+    if (modeWatch === "absorbance") {
+      setValue("simulate_slit", undefined);
+    } else {
+      setValue("simulate_slit", 5);
+    }
+  }, [databaseWatch, modeWatch]);
+
+  const handleBadResponse = (message: string) => {
+    setCalcSpectrumResponse(undefined);
+    setError(message);
+  };
+  const onSubmit = async (
+    data: FormValues,
+    endpoint: string
+  ): Promise<void> => {
+    if (useSlit == true) {
+      if (data.mode === "radiance_noslit") {
+        data.mode = "radiance";
+      }
+      if (data.mode === "transmittance_noslit") {
+        data.mode = "transmittance";
+      }
+    }
+
+    const molecules = data.species.map(({ molecule }) => molecule).join("_");
+
+    setDownloadButton(true);
+    setLoading(true);
+    setError(undefined);
+    setPlotData({
+      max_wavenumber_range: data.max_wavenumber_range,
+      min_wavenumber_range: data.min_wavenumber_range,
+      mode: data.mode,
+      species: data.species,
+    });
+
+    import(/* webpackIgnore: true */ "./config.js").then(async (module) => {
+      if (endpoint === "calculate-spectrum") {
+        setProgress(30);
+
+        const rawResponse = await axios({
+          url: module.apiEndpoint + `calculate-spectrum`,
+          method: "POST",
+          data: data,
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+        if (
+          rawResponse.data.data === undefined &&
+          rawResponse.data.error === undefined
+        ) {
+          handleBadResponse("Bad response from backend!");
+          setDownloadButton(true);
+        } else {
+          const response = await rawResponse.data;
+          if (response.error) {
+            handleBadResponse(response.error);
+            setDownloadButton(true);
+          } else {
+            setCalcSpectrumResponse(response);
+            setDownloadButton(false);
+          }
+        }
+
+        setProgress(100);
+        setLoading(false);
+      }
+
+      if (endpoint === "download-spectrum") {
+        setProgress(30);
+        setLoading(false);
+        const rawResponse = await axios({
+          url: module.apiEndpoint + `download-spectrum`,
+          method: "POST",
+          responseType: "blob",
+          data: data,
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+        const url = window.URL.createObjectURL(new Blob([rawResponse.data]));
+        const link = document.createElement("a");
+        link.href = url;
+        link.setAttribute(
+          "download",
+          `${data.database}_${molecules}_${data.min_wavenumber_range}_${data.max_wavenumber_range}cm-1_${data.tgas}K_${data.pressure}atm.spec`
+        );
+        document.body.appendChild(link);
+        link.click();
+        setDownloadButton(false);
+        const response = await rawResponse.data;
+        if (response.error) {
+          handleBadResponse(response.error);
+        } else {
+          setDownloadButton(false);
+        }
+        setDownloadButton(false);
+        setProgress(100);
+      }
+    });
+  };
+
   //equilibrium-switch
   const UseNonEquilibriumCalculations = () => (
     <Controller
@@ -121,80 +330,85 @@ export const Form: React.FunctionComponent<FormProps> = ({
       Download
     </Button>
   );
+
   return (
-    <Grid container spacing={3}>
-      <Grid item xs={12} sm={8} md={5} lg={5}>
-        <Database control={control}></Database>
-      </Grid>
-      <Grid item xs={12} sm={8} md={5} lg={6}>
-        <Mode control={control} />
-      </Grid>
-      <Grid item xs={12}>
-        <WavenumberRangeSlider
-          minRange={500}
-          maxRange={10000}
-          control={control}
-          setValue={setValue}
-        />
-      </Grid>
-
-      <Grid item sm={8} lg={4}>
-        <TGas control={control} />
-      </Grid>
-
-      {isNonEquilibrium ? (
-        <>
-          <Grid item sm={8} lg={3}>
-            <TRot control={control} />
-          </Grid>
-          <Grid item sm={8} lg={3}>
-            <TVib control={control} />
-          </Grid>
-        </>
-      ) : null}
-
-      <Grid item sm={8} lg={5}>
-        <Pressure control={control} />
-      </Grid>
-
-      <Grid item sm={8} lg={3}>
-        <PathLength control={control} />
-      </Grid>
-
-      <Grid item xs={12}>
-        <Species
-          isNonEquilibrium={isNonEquilibrium}
-          control={control}
-          isGeisa={useGesia}
-          isHitemp={useHitemp}
-        />
-      </Grid>
-
-      {useSimulateSlitFunction ? (
-        <Grid item xs={12}>
-          <UseSimulateSlit />
+    <form
+      onSubmit={handleSubmit((data) => onSubmit(data, `calculate-spectrum`))}
+    >
+      <Grid container spacing={3}>
+        <Grid item xs={12} sm={8} md={5} lg={5}>
+          <Database control={control}></Database>
         </Grid>
-      ) : null}
+        <Grid item xs={12} sm={8} md={5} lg={6}>
+          <Mode control={control} />
+        </Grid>
+        <Grid item xs={12}>
+          <WavenumberRangeSlider
+            minRange={500}
+            maxRange={10000}
+            control={control}
+            setValue={setValue}
+          />
+        </Grid>
 
-      {useSimulateSlitFunction ? (
-        useSlit ? (
+        <Grid item sm={8} lg={4}>
+          <TGas control={control} />
+        </Grid>
+
+        {isNonEquilibrium ? (
+          <>
+            <Grid item sm={8} lg={3}>
+              <TRot control={control} />
+            </Grid>
+            <Grid item sm={8} lg={3}>
+              <TVib control={control} />
+            </Grid>
+          </>
+        ) : null}
+
+        <Grid item sm={8} lg={5}>
+          <Pressure control={control} />
+        </Grid>
+
+        <Grid item sm={8} lg={3}>
+          <PathLength control={control} />
+        </Grid>
+
+        <Grid item xs={12}>
+          <Species
+            isNonEquilibrium={isNonEquilibrium}
+            control={control}
+            isGeisa={useGesia}
+            isHitemp={useHitemp}
+          />
+        </Grid>
+
+        {useSimulateSlitFunction ? (
           <Grid item xs={12}>
-            <SimulateSlit control={control} />
+            <UseSimulateSlit />
           </Grid>
-        ) : null
-      ) : null}
-      {useGesia ? null : (
-        <Grid item xs={12}>
-          <UseNonEquilibriumCalculations />
-        </Grid>
-      )}
+        ) : null}
 
-      <Grid item xs={12}>
-        <CalcSpectrumButton />
+        {useSimulateSlitFunction ? (
+          useSlit ? (
+            <Grid item xs={12}>
+              <SimulateSlit control={control} />
+            </Grid>
+          ) : null
+        ) : null}
+        {useGesia ? null : (
+          <Grid item xs={12}>
+            <UseNonEquilibriumCalculations />
+          </Grid>
+        )}
+
+        <Grid item xs={12}>
+          <CalcSpectrumButton />
+        </Grid>
+        <Grid item xs={12}>
+          <DownloadSpectrum />
+        </Grid>
       </Grid>
-      <Grid item xs={12}>
-        <DownloadSpectrum />
-      </Grid>
-    </Grid>
+    </form>
   );
 };
